requirements.txt
requests>=2.32.0
requests-oauthlib>=2.0.0
Flask>=3.0.0
python-dotenv>=1.0.1
pyjwt>=2.9.0
cryptography>=43.0.0
